chore: rename utils directory to ui_utils
