+++ frontend/src/js/ui_utils/hotbar_utils.js
