+++ frontend/src/js/ui_utils/input_field_utils.js
